clubApi 구현

+++ src/main/java/stg/onyou/service/FeedService.java
@@ -20,7 +20,7 @@ public class FeedService {
      * feed upload
      */
 
-    public Long upload(Feed feed) {
+    public int upload(Feed feed) {
         feedRepository.save(feed);
         return feed.getId();
     }
